scripts/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
